// backend/src/types/AcademicDocumentType.ts
export enum AcademicDocumentType {
    BIRTH_CERTIFICATE = 'BIRTH_CERTIFICATE',
    AADHAR_CARD = 'AADHAR_CARD',
    MARKSHEET_8 = 'MARKSHEET_8',
    MARKSHEET_9 = 'MARKSHEET_9',
    MARKSHEET_10 = 'MARKSHEET_10',
    MARKSHEET_11 = 'MARKSHEET_11',
    MARKSHEET_12 = 'MARKSHEET_12',
    UNIVERSITY_MARKSHEET = 'UNIVERSITY_MARKSHEET',
}